Introduce kitchen-terraform tests
This commit includes all changes needed to implement kitchen-terraform
tests for the log-exports module. Tests are structured exactly the same
as examples: folders for each of the destinations, and subfolders for
each log sink level that's possible for each destination (with
billing_account as an exception - more on that later). Each test fixture
functions independently and relies upon environment variables that are
set within `test/ci_integration.sh` (environment variables for the SA
key json, project ID, and the parent resource
project/folder/org/billing_account). Like all the other CFT modules,
tests are invoked based on make targets from `Makefile`, and there are
dedicated targets for testing from the CFT Docker container as well as
directly from your laptop/machine.

A test fixture called `computed_values` has also been provided to
specifically test issue #10 of the log-export module. This test spins up
a project on the fly and passes that computed project ID to the
log-export module to ensure that the module can handle computed values.

Tests for the billing account level log sinks are not provided within this
commit and will be provided in a later commit/PR.